refactor: Move test module
